finish Step3
