+++ src/Step1/MapList.java
@@ -56,8 +56,8 @@ public class MapList {
                 {"#", " ", " ", " ", "#", " ", " ", " ", "#"},
                 {"#", " ", " ", "o", "o", " ", "o", " ", "#"},
                 {"#", " ", " ", " ", "#", "P", "#", " ", "#"},
-                {"#", " ", " ", " ", "#", "o", "#", " ", "#"},
-                {"#", " ", " ", " ", "O", "O", "O", "O", "#"},
+                {"#", " ", " ", " ", "#", " ", "#", " ", "#"},
+                {"#", " ", " ", " ", "O", " ", "O", "O", "#"},
                 {"#", " ", " ", " ", "#", " ", " ", "#", "#"},
                 {"#", "#", "#", "#", "#", "#", "#", "#", "#"}
         };
